refactor: reorganize and extract CatchupIter and MaybeBlock to client module
