+++ chain-signatures/chain-ethereum/src/indexer_eth_direct_rpc.rs
@@ -1,4 +1,3 @@
-use crate::indexer::MaybeBlock;
 use alloy::eips::BlockNumberOrTag;
 use alloy::primitives::hex::{self, ToHexExt};
 use alloy::primitives::{Address, Bytes, B256};
@@ -9,6 +8,8 @@ use serde_json::json;
 use std::collections::HashMap;
 use std::sync::atomic::{AtomicU64, Ordering};
 use std::sync::Arc;
+
+use crate::client::MaybeBlock;
 
 // This is more than likely limited by the RPC provider, but alchemy
 // supports archive nodes, so we effectively can go as far back as needed
@@ -365,7 +366,6 @@ fn to_hex_block_id(block_id: BlockId) -> String {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::indexer::MaybeBlock;
     use alloy::eips::BlockNumberOrTag;
     use alloy::primitives::B256;
     use alloy::rpc::types::BlockId;
